Add files via upload
This project focuses on analyzing data about fictional Android and iOS apps that are made available on Google Play and in the App Store.

All apps are made free to download and the only source of revenue comes from in-app ads. Therefore, the number of users determines the profit from each app: the more engaging the add, the more people will use the app.